+++ mrerr/field_error_list.go
@@ -12,8 +12,8 @@ func (l *FieldErrorList) Add(fieldID string, err error) {
 	*l = append(*l, NewFieldError(fieldID, err))
 }
 
-func (l *FieldErrorList) AddAppErr(fieldID string, err *AppError) {
-	*l = append(*l, NewFieldErrorAppErr(fieldID, err))
+func (l *FieldErrorList) AddAppError(fieldID string, err *AppError) {
+	*l = append(*l, NewFieldErrorAppError(fieldID, err))
 }
 
 func (l FieldErrorList) Error() string {